Fix for Derby - 67 .  Network Server on a 64 bit JVM failes with Distributed Protocol Error
git-svn-id: https://svn.apache.org/repos/asf/incubator/derby/code/trunk@106935 13f79535-47bb-0310-9956-ffa450edef68

+++ java/drda/org/apache/derby/impl/drda/DRDAConnThread.java
@@ -6913,7 +6913,7 @@ public class DRDAConnThread extends Thread {
 	private boolean readBoolean(int codepoint) throws DRDAProtocolException
 	{
 		checkLength(codepoint, 1);
-		int val = reader.readByte();
+		byte val = reader.readByte();
 		if (val == CodePoint.TRUE)
 			return true;
 		else if (val == CodePoint.FALSE)
